Added some checks to conf

+++ conf.sh
@@ -6,6 +6,8 @@ if [ "$response" != "y" ]; then
     exit 1
 fi
 
+mkdir ~/.config
+
 for folder in $(ls ./config); do
     rm -rf ~/.config/$folder
 done
@@ -22,17 +24,19 @@ for folder in $(ls ./etc); do
     sudo cp -r $(pwd)/etc/$folder /etc/$folder
 done
 
+sudo pacman -S git base-devel
+
+if pacman -Q yay &> /dev/null; then
+    git clone https://aur.archlinux.org/yay.git
+    cd yay
+    makepkg -si
+    cd ..
+    rm -rf yay
+fi
+
+yay -S $(cat ./pkglist)
+
 sh -c "$(wget https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh -O -)"
 
 cp ./home/zshrc ~/.zshrc
 cp ./home/xinitrc ~/.xinitrc
-
-# Clone yay and install it
-git clone https://aur.archlinux.org/yay.git
-cd yay
-makepkg -si
-cd ..
-rm -rf yay
-
-# Install packages
-yay -S $(cat ./pkglist)
